# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a group of API for WTM Service",
        "title": "WTM Backend Service",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/email/contact-us": {
            "post": {
                "description": "Send a contact us email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Send Contact Us Email",
                "parameters": [
                    {
                        "description": "Contact Us Email Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emaildto.SendContactUsEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully sent contact us email",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/hotels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of hotels using query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "List Hotels",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter hotels by API status",
                        "name": "is_api",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by region",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter hotels by status_id",
                        "name": "status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword to filter hotels by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of hotels",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/hoteldto.ListHotel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "Create Hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Hotel photos (multiple allowed)",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sub-district location",
                        "name": "sub_district",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "District location",
                        "name": "district",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Province location",
                        "name": "province",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Hotel rating (1–5)",
                        "name": "rating",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nearby places as JSON string. Example: /example_nearby_places ",
                        "name": "nearby_places",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Facilities (multiple allowed)",
                        "name": "facilities",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Social media links as JSON string. Example: /example_social_medias ",
                        "name": "social_medias",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created hotel",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hoteldto.CreateHotelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/hotels/room-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of room types for a specific hotel.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "List Room Types",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hotel Id to filter room types",
                        "name": "hotel_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of room types",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/hoteldto.ListRoomType"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new room type to a hotel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "Add Room Type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hotel ID",
                        "name": "hotel_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Room type name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Room type photos (multiple allowed)",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Without breakfast details as JSON string. Example : /example_room_options?room_options=without_breakfast",
                        "name": "without_breakfast",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "With breakfast details as JSON string. Example : /example_room_options?room_options=with_breakfast",
                        "name": "with_breakfast",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Room size in square meters",
                        "name": "room_size",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum occupancy",
                        "name": "max_occupancy",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Bed types (multiple allowed)",
                        "name": "bed_types",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is smoking room",
                        "name": "is_smoking_room",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Additional room features as JSON string. Example: /example_additional_features",
                        "name": "additional",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Room type description",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added room type",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emaildto.SendContactUsEmailRequest": {
            "type": "object",
            "properties": {
                "booking_code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sub_booking_code": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "hoteldto.CreateHotelResponse": {
            "type": "object",
            "properties": {
                "hotel_id": {
                    "type": "integer"
                }
            }
        },
...
        "hoteldto.ListHotel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_api": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hoteldto.RoomTypeItem"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "hoteldto.ListRoomType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "hoteldto.RoomTypeItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "price_with_breakfast": {
                    "type": "number"
                }
            }
        },
...
        "response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ResponseWithPagination": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}